pass -Wno-error to preprocessor when run_second_cpp = true
clang issues warnings on unused command-line arguments. When only
running the pre-processor, options that affect linking, assembly and
code generation are reported as unused. When -Werror is passed during
compilation, these warnings are turned into errors and ccache fails
to successfully run the preprocessor.

This patch scans through the command-line arguments and checks for the
presence of -Werror, and, if found adds a -Wno-error to the preprocessor
command line arguments. This is only done when running with
run_second_cpp = true, as otherwise errors might not get reported.